+++ housekeeper/store/api/base.py
@@ -1,0 +1,9 @@
+from housekeeper.store import models
+
+
+class BaseHandler:
+
+    Bundle = models.Bundle
+    Version = models.Version
+    File = models.File
+    Tag = models.Tag
